feat: add 'register' parameter to Item constructor
Introduced a new 'register' parameter in the Item constructor within IItemProps. This parameter allows specification of whether collections subscribed to the defined entity need to be informed about this item. It is particularly useful for creating items that already exist outside the application's memory. If 'register' is set to true, the registry will trigger an event when the item is created to notify the collections.